Refactor out the duplicate definition of method `isEntireColumn`
